// web/admin/gql/schema.tsx
export const crs = { type: 'name', properties: { name: 'urn:ogc:def:crs:EPSG::4326' } }
export type Maybe<T> = T | null | undefined
// ...
export enum SpaceCategory {
  Private = 'PRIVATE',
  Public = 'PUBLIC',
}

/** Allowed statuses are: enabled, disabled, pending */
export enum ParkingSpaceStatus {
  Enabled = 'ENABLED',
  Disabled = 'DISABLED',
  Pending = 'PENDING',
  Unlisted = 'UNLISTED',
  Deleted = 'DELETED',
}

/** Allowed values for space_verification_status are: pending, rejected, verified */
export enum SpaceVerificationStatus {
  Pending = 'PENDING',
  Rejected = 'REJECTED',
  Verified = 'VERIFIED',
}

/** Allowed values for access_restrictions are: none, barrier, keycard, keycode */
export enum SpaceAccessRestriction {
  None = 'NONE',
  Barrier = 'BARRIER',
  Keycard = 'KEYCARD',
  Keycode = 'KEYCODE',
}
// ...
export type CreateParkingSpaceMutation = {
  createParkingSpace?: Maybe<{
    parkingSpace?: Maybe<{
      id: string
      name: string
      slug?: Maybe<string>
      address?: Maybe<{ [key: string]: unknown }>
      status?: Maybe<ParkingSpaceStatus>
      location?: Maybe<{ longitude: number; latitude: number }>
      carEntry?: Maybe<{ longitude: number; latitude: number }>
      carExit?: Maybe<{ longitude: number; latitude: number }>
      parkingSpaceAvailabilitiesList: Array<{
        id: string
        parkingSpaceId: string
        fromDate: any
        toDate: any
        defaultFlag: boolean
        closedFlag: boolean
        parkingWorkingHoursList: Array<{
          id: string
          parkingSpaceAvailabilityId: string
          dayOfWeek: { [key: string]: unknown }
          fromTime: string
          toTime: string
        }>
        parkingOpenHoursList: Array<{
          id: string
          parkingSpaceAvailabilityId: string
          dayOfWeek: { [key: string]: unknown }
          fromTime: string
          toTime: string
          price: number
          currency: string
        }>
      }>
    }>
  }>
}

export type DeleteParkingSpaceMutationVariables = Exact<{
  id: Scalars['UUID']
}>

export type DeleteParkingSpaceMutation = {
  deleteParkingSpace?: Maybe<{ deletedParkingSpaceNodeId?: Maybe<string> }>
}

export type UpdateParkingSpaceMutationVariables = Exact<{
  id: Scalars['UUID']
  patch: ParkingSpacePatch
}>

export type UpdateParkingSpaceMutation = {
  updateParkingSpace?: Maybe<{
    parkingSpace?: Maybe<{
      id: string
      name: string
      slug?: Maybe<string>
      address?: Maybe<{ [key: string]: unknown }>
      status?: Maybe<ParkingSpaceStatus>
      location?: Maybe<{ longitude: number; latitude: number }>
      carEntry?: Maybe<{ longitude: number; latitude: number }>
      carExit?: Maybe<{ longitude: number; latitude: number }>
    }>
  }>
}
// ...
export type ManageSpaceQuery = {
  parkingSpace?: Maybe<{
    id: string
    ownerId: string
    name: string
    description?: Maybe<string>
    photoUrl?: Maybe<string>
    address?: Maybe<{ [key: string]: unknown }>
    languageCode?: Maybe<string>
    floor?: Maybe<number>
    category?: Maybe<SpaceCategory>
    status?: Maybe<ParkingSpaceStatus>
    accessRestriction?: Maybe<Array<Maybe<SpaceAccessRestriction>>>
    businessStatusReason?: Maybe<string>
    location?: Maybe<{ latitude: number; longitude: number }>
    carEntry?: Maybe<{ x: number; y: number }>
    carExit?: Maybe<{ x: number; y: number }>
    companyEntrance?: Maybe<{ x: number; y: number }>
    parkingspaceMapview?: Maybe<{ geojson?: Maybe<GeometryObject> }>
    parkingSpaceAvailabilitiesList: Array<{
      fromDate: any
      toDate: any
      defaultFlag: boolean
      closedFlag: boolean
      parkingWorkingHoursList: Array<{
        dayOfWeek: { [key: string]: unknown }
        fromTime: string
        toTime: string
      }>
      parkingOpenHoursList: Array<{
        dayOfWeek: { [key: string]: unknown }
        fromTime: string
        toTime: string
        price: number
        currency?: Maybe<string>
      }>
    }>
  }>
}